third_party/WebKit/LayoutTests/webmidi/loopback-receive-expected.txt

Test if midimessage event handler work.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.requestMIDIAccess is defined.
running test with useAddEventListener = false
receiving data 0: 128,0,0
PASS actual is expected
receiving data 1: 144,0,0
PASS actual is expected
receiving data 2: 160,0,0
PASS actual is expected
receiving data 3: 176,0,0
PASS actual is expected
receiving data 4: 192,0
PASS actual is expected
receiving data 5: 208,0
PASS actual is expected
receiving data 6: 224,0,0
PASS actual is expected
receiving data 7: 248
PASS actual is expected
receiving data 8: 250
PASS actual is expected
receiving data 9: 251
PASS actual is expected
receiving data 10: 252
PASS actual is expected
receiving data 11: 254
PASS actual is expected
receiving data 12: 255
PASS actual is expected
receiving data 13: 240,0,1,2,3,247
PASS actual is expected
receiving data 14: 240,248,247
PASS actual is expected
running test with useAddEventListener = true
receiving data 0: 128,0,0
PASS actual is expected
receiving data 1: 144,0,0
PASS actual is expected
receiving data 2: 160,0,0
PASS actual is expected
receiving data 3: 176,0,0
PASS actual is expected
receiving data 4: 192,0
PASS actual is expected
receiving data 5: 208,0
PASS actual is expected
receiving data 6: 224,0,0
PASS actual is expected
receiving data 7: 248
PASS actual is expected
receiving data 8: 250
PASS actual is expected
receiving data 9: 251
PASS actual is expected
receiving data 10: 252
PASS actual is expected
receiving data 11: 254
PASS actual is expected
receiving data 12: 255
PASS actual is expected
receiving data 13: 240,0,1,2,3,247
PASS actual is expected
receiving data 14: 240,248,247
PASS actual is expected
PASS successfullyParsed is true

TEST COMPLETE
